Refactor + rename files
