+++ examples/lexicons/codelingo/ast/typescript/src/babel/rename.ts
