+++ examples/lexicons/codelingo/ast/typescript/src/common/keys.ts
